feat: Dynamic RelationTypes, JSON FoundationConcepts, ConceptEmbeddingStore, GlobalDynamicsOperator
- RelationTypeRegistry: dynamic relation types with categories (replaces static enum)
- FoundationConcepts: loaded from JSON config instead of hardcoded
- ConceptEmbeddingStore: embedding management without Ollama
- GlobalDynamicsOperator: background activation landscape
- All modules tested and compiling

+++ backend/persistent/persistent_ltm.hpp
@@ -73,7 +73,7 @@ public:
     );
     void replay_add_relation(
         uint64_t relation_id, uint64_t source, uint64_t target,
-        uint8_t type, double weight
+        uint16_t type, double weight
     );
     
 private:
